requirements.txt
streamlit==1.45.0
numpy==2.1.3
tensorflow==2.19.0
pillow==11.2.1
